Extract works when not using multiple workers.
